Adjust some of the scraping logic and data exposed
Switch to fs.promises to simplify syntax for async fs operations.

Attempt to load the existing data file prior to scraping, in
order to avoid re-running a geocoding operation against an
address that's already been coded and locally available.

Attempt to find site type headers and collect additional information
from them to include with the related entries, otherwise rely on the
original element identification logic.

Expose the site type and testing type in the pop-up.